Add Uses Texture Mapping Method
In some cases, it may be desirable to have a texture node which doesn't use texture mapping.
Added support for this in the code by replacing some calls to is texture node to uses texture mapping.
Modified/added tests to support this feature.
Added vector socket support for OSL paramater