Refactor GLSL
Pass the gui object directly to the glsl manager to reduce the length of the constructor
parameter list. As it was, creating new features which require additional info would mean extending
the constructor, and rewrite existing code. Assuming all required info is accessible through the gui object,
which is a reasonable assumption to make, this will eliminate that issue.
Refactored tests for this change.